pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "discord-py>=2.6.4",
    "nest-asyncio>=1.6.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
